Updates 14/12/2023
[DONE]

- Country
- Checkbox component

...............................................................................
[PENDING]

- Product
- Company
- Address with City

+++ routes/web.php
@@ -18,6 +18,11 @@ use App\Http\Controllers\Role\CRUDController as RoleCRUDController;
 
 // Permission Controllers
 use App\Http\Controllers\Permission\CRUDController as PermissionCRUDController;
+
+// Country Controllers
+use App\Http\Controllers\Country\CRUDController as CountryCRUDController;
+use App\Http\Middleware\EnsureCountryExists;
+
 use App\Http\Middleware\EnsureCategoryExists;
 use App\Http\Middleware\EnsureRoleExists;
 use App\Http\Middleware\EnsurePermissionExists;
@@ -30,6 +35,9 @@ use App\Http\Middleware\RedirectIfAuthenticated;
 use App\Models\Category;
 use App\Models\Role;
 use App\Models\Permission;
+use App\Models\Country;
+
+
 
 /*
 |--------------------------------------------------------------------------
@@ -109,11 +117,11 @@ Route::name('admin.')->prefix('/admin')->group(function () {
   Route::middleware([])->group(function () {
     Route::get('/categories', [CategoryCRUDController::class, 'index'])->name('categories.index');
     Route::name('category.')->prefix('/categories')->group(function () {
-      Route::get('/create', [CategoryCRUDController::class, 'create'])->name('create')->can('create', Category::class);
-      Route::post('/store', [CategoryCRUDController::class, 'store'])->name('store');
-      Route::get('/{category}/edit', [CategoryCRUDController::class, 'edit'])->middleware(EnsureCategoryExists::class)->name('edit');
-      Route::patch('/{category}/update', [CategoryCRUDController::class, 'update'])->middleware(EnsureCategoryExists::class)->name('update');
-      Route::delete('/{category}/destroy', [CategoryCRUDController::class, 'destroy'])->name('destroy');
+      Route::get('/create', [CategoryCRUDController::class, 'create'])->name('create')->can('create category', Category::class);
+      Route::post('/store', [CategoryCRUDController::class, 'store'])->name('store')->can('create category', Category::class);
+      Route::get('/{category}/edit', [CategoryCRUDController::class, 'edit'])->middleware(EnsureCategoryExists::class)->name('edit')->can('update category', Category::class);
+      Route::patch('/{category}/update', [CategoryCRUDController::class, 'update'])->middleware(EnsureCategoryExists::class)->name('update')->can('update category', Category::class);
+      Route::delete('/{category}/destroy', [CategoryCRUDController::class, 'destroy'])->name('destroy')->can('delete category', Category::class);
     });
   });
   // Role Routes (CRUD)
@@ -121,10 +129,10 @@ Route::name('admin.')->prefix('/admin')->group(function () {
     Route::get('/roles', [RoleCRUDController::class, 'index'])->name('roles.index');
     Route::name('role.')->prefix('/roles')->group(function () {
       Route::get('/create', [RoleCRUDController::class, 'create'])->name('create')->can('create', Role::class);
-      Route::post('/store', [RoleCRUDController::class, 'store'])->name('store');
-      Route::get('/{role}/edit', [RoleCRUDController::class, 'edit'])->middleware([EnsureRoleExists::class, EnsureRoleIsNotLocked::class])->name('edit');
-      Route::patch('/{role}/update', [RoleCRUDController::class, 'update'])->middleware(EnsureRoleExists::class, EnsureRoleIsNotLocked::class)->name('update');
-      Route::delete('/{role}/destroy', [RoleCRUDController::class, 'destroy'])->name('destroy');
+      Route::post('/store', [RoleCRUDController::class, 'store'])->name('store')->can('create', Role::class);
+      Route::get('/{role}/edit', [RoleCRUDController::class, 'edit'])->middleware([EnsureRoleExists::class, EnsureRoleIsNotLocked::class])->name('edit')->can('update', Role::class);
+      Route::patch('/{role}/update', [RoleCRUDController::class, 'update'])->middleware(EnsureRoleExists::class, EnsureRoleIsNotLocked::class)->name('update')->can('update', Role::class);
+      Route::delete('/{role}/destroy', [RoleCRUDController::class, 'destroy'])->name('destroy')->can('delete', Role::class);
     });
   });
   // Permission Routes (CRUD)
@@ -132,10 +140,18 @@ Route::name('admin.')->prefix('/admin')->group(function () {
     Route::get('/permissions', [PermissionCRUDController::class, 'index'])->name('permissions.index');
     Route::name('permission.')->prefix('/permissions')->group(function () {
       Route::get('/create', [PermissionCRUDController::class, 'create'])->name('create')->can('create', Permission::class);
-      Route::post('/store', [PermissionCRUDController::class, 'store'])->name('store');
-      Route::get('/{permission}/edit', [PermissionCRUDController::class, 'edit'])->middleware([EnsurePermissionExists::class, EnsurePermissionIsNotLocked::class])->name('edit');
-      Route::patch('/{permission}/update', [PermissionCRUDController::class, 'update'])->middleware([EnsurePermissionExists::class, EnsurePermissionIsNotLocked::class])->name('update');
-      Route::delete('/{permission}/destroy', [PermissionCRUDController::class, 'destroy'])->name('destroy');
+      Route::post('/store', [PermissionCRUDController::class, 'store'])->name('store')->can('create', Permission::class);
+      Route::get('/{permission}/edit', [PermissionCRUDController::class, 'edit'])->middleware([EnsurePermissionExists::class, EnsurePermissionIsNotLocked::class])->name('edit')->can('update', Permission::class);
+      Route::patch('/{permission}/update', [PermissionCRUDController::class, 'update'])->middleware([EnsurePermissionExists::class, EnsurePermissionIsNotLocked::class])->name('update')->can('update', Permission::class);
+      Route::delete('/{permission}/destroy', [PermissionCRUDController::class, 'destroy'])->name('destroy')->can('delete', Permission::class);
+    });
+  });
+  // Country Routes (CRUD)
+  Route::middleware([])->group(function () {
+    Route::get('/countries', [CountryCRUDController::class, 'index'])->name('countries.index');
+    Route::name('country.')->prefix('/countries')->group(function () {
+      Route::get('/{country}/edit', [CountryCRUDController::class, 'edit'])->middleware([EnsureCountryExists::class])->name('edit')->can('update', Country::class);
+      Route::patch('/{country}/update', [CountryCRUDController::class, 'update'])->middleware([EnsureCountryExists::class])->name('update')->can('update', Country::class);
     });
   });
 });
